refactor: move creation of main header and tab menu to Top.js
- Rename constants of main header and tab menu inner text
- Move creation of main header and tab menu to Top.js
