+++ src/constants.js
@@ -3,7 +3,7 @@ export const HEADER = {
 };
 
 export const MENU = {
-  PRODUCT_ADD: '상품 관리',
-  VENDING_MACHINE_MANAGE: '잔돈 충전',
+  PRODUCT_MANAGE: '상품 관리',
+  CHANGE_CHARGE: '잔돈 충전',
   PRODUCT_PURCHASE: '상품 구매',
 };
